+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-# [8.14.0] - 2024-11-??
+# [8.14.0] - 2024-11-14
 - Close HTTP responses to prevent resource leaks
 - Added `RedactResponseException` and deprecated `VonageBadRequestException`
 - Added `maxBitrate` to `com.vonage.client.video.Archive`
+- Added varargs `removeCapabilities` method to `Application.Builder`
 
 # [8.13.0] - 2024-10-28
 - Added support for Verify custom templates